+++ tests/testthat/test-rules.R
@@ -41,9 +41,9 @@ test_that("rules works with `dplyr` version higher than 0.8.0",  {
 
   output_1 <- rules(mean, "mean", mean(.), ~ mean(.))
   output_ref_1 <- list(
-    ._.rule__1 = rlang::quo(mean),
-    ._.rule__2 = rlang::quo("mean"),
-    ._.rule__3 = rlang::quo(mean(.)),
+    ._.rule__1 = mean,
+    ._.rule__2 = "mean",
+    ._.rule__3 = ~ mean(.),
     ._.rule__4 = output_1[[4]]
   )
 
@@ -53,4 +53,22 @@ test_that("rules works with `dplyr` version higher than 0.8.0",  {
   output_ref_2 <- list(a_a_rule__1 = output_2[[1]])
 
   expect_identical(output_2, output_ref_2)
+
+  expect_error(rules(mean2), "`mean2`")
 })
+
+
+# extract_funs_input ------------------------------------------------------
+# Tested in `rules()`
+
+
+# has_dot_symbol ----------------------------------------------------------
+# Tested in `rules()`
+
+
+# squash_expr -------------------------------------------------------------
+# Tested in `rules()`
+
+
+# quo_get_function --------------------------------------------------------
+# Tested in `rules()`